# specs/001-todo-frontend-auth/contracts/api.yaml
openapi: 3.0.0
info:
  title: Todo Frontend API Contracts
  version: 1.0.0
  description: API contracts for frontend-backend communication in the Todo application

paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
      responses:
        '200':
          description: List of user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in token doesn't match path
      security:
        - bearerAuth: []

    post:
      summary: Create a new task for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in token doesn't match path
      security:
        - bearerAuth: []

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in token doesn't match path or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

    put:
      summary: Update a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in token doesn't match path or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

    delete:
      summary: Delete a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in token doesn't match path or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: The new completion status
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in token doesn't match path or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the task
        title:
          type: string
          description: Title of the task
        description:
          type: string
          description: Optional description of the task
        completed:
          type: boolean
          description: Whether the task is completed
        userId:
          type: string
          description: ID of the user who owns the task
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the task was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
      required:
        - id
        - title
        - completed
        - userId
        - createdAt
        - updatedAt

    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the task (required)
        description:
          type: string
          description: Optional description of the task
      required:
        - title

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the task
        description:
          type: string
          description: Updated description of the task
        completed:
          type: boolean
          description: Updated completion status

security:
  - bearerAuth: []